Correction of comments
